+++ test/test_run_experiments/test_run_mysql_experiment.py
@@ -35,7 +35,7 @@ def check_done_entries(experimenter: PyExperimenter, amount_of_entries: int):
 def test_run_all_mqsql_experiments():
     experiment_configuration_file_path = os.path.join("test", "test_run_experiments", "test_run_mysql_experiment_config.yml")
     logging.basicConfig(level=logging.DEBUG)
-    experimenter = PyExperimenter(experiment_configuration_file_path=experiment_configuration_file_path, use_codecarbon=False)
+    experimenter = PyExperimenter(experiment_configuration_file_path=experiment_configuration_file_path, use_codecarbon=False, use_ssh_tunnel=False)
     try:
         experimenter.delete_table()
     except ProgrammingError as e:
@@ -52,8 +52,11 @@ def test_run_all_mqsql_experiments():
     entries_without_metadata = entries[0][:3] + (entries[0][4],) + entries[0][6:7] + entries[0][8:10] + (entries[0][-1],)
     assert entries_without_metadata == (1, 1, 1, "done", "PyExperimenter", 0.841471, 0.540302, None)
     experimenter.db_connector.close_connection(connection)
+    experimenter.close_ssh()
+    cursor = None
+    connection = None
 
-    experimenter = PyExperimenter(experiment_configuration_file_path=experiment_configuration_file_path, use_codecarbon=False)
+    experimenter = PyExperimenter(experiment_configuration_file_path=experiment_configuration_file_path, use_codecarbon=False, use_ssh_tunnel=False)
     experimenter.fill_table_from_config()
     experimenter.execute(own_function, max_experiments=-1)
     check_done_entries(experimenter, 30)
@@ -77,6 +80,29 @@ def test_run_all_mqsql_experiments():
     assert set(range(2, 32)) == set(entry[0] for entry in entries)
 
 
+def test_mysql_shh():
+    experiment_configuration_file_path = os.path.join("test", "test_run_experiments", "test_run_mysql_experiment_config.yml")
+    logging.basicConfig(level=logging.DEBUG)
+    experimenter = PyExperimenter(experiment_configuration_file_path=experiment_configuration_file_path, use_codecarbon=False, use_ssh_tunnel=True)
+    try:
+        experimenter.delete_table()
+    except ProgrammingError as e:
+        logging.warning(e)
+    experimenter.fill_table_from_config()
+    experimenter.execute(own_function, max_experiments=1)
+
+    connection = experimenter.db_connector.connect()
+    cursor = experimenter.db_connector.cursor(connection)
+    cursor.execute(f"SELECT * FROM {experimenter.db_connector.database_configuration.table_name} WHERE status = 'done'")
+    entries = cursor.fetchall()
+
+    assert len(entries) == 1
+    entries_without_metadata = entries[0][:3] + (entries[0][4],) + entries[0][6:7] + entries[0][8:10] + (entries[0][-1],)
+    assert entries_without_metadata == (1, 1, 1, "done", "PyExperimenter", 0.841471, 0.540302, None)
+    experimenter.db_connector.close_connection(connection)
+    experimenter.close_ssh()
+
+
 def error_function(keyfields: dict, result_processor: ResultProcessor, custom_fields: dict):
     raise Exception("Error with weird symbos '@#$%&/\()=")
 
@@ -93,7 +119,7 @@ def check_error_entries(experimenter: PyExperimenter):
 def test_run_error_experiment():
     experiment_configuration_file_path = os.path.join("test", "test_run_experiments", "test_run_mysql_experiment_config.yml")
     logging.basicConfig(level=logging.DEBUG)
-    experimenter = PyExperimenter(experiment_configuration_file_path=experiment_configuration_file_path, use_codecarbon=False)
+    experimenter = PyExperimenter(experiment_configuration_file_path=experiment_configuration_file_path, use_codecarbon=False, use_ssh_tunnel=False)
     try:
         experimenter.delete_table()
     except ProgrammingError as e:
@@ -114,16 +140,3 @@ def test_run_error_experiment():
         "Error with weird symbos '@#$%&/\\()=",
     ]:
         assert message in entries[0][11]
-
-
-def own_function_raising_errors(keyfields: dict, result_processor: ResultProcessor, custom_fields: dict):
-    error_code = keyfields["error_code"]
-
-    # give list of special characters that frequently cause problems
-    characters = ['"', "'", "@", "#", "$", "%", "&", "/", "\\", "(", ")", "=", "`", "`some_text`", "^"]
-    if error_code == 0:
-        raise Exception("Error with weird symbos" + "".join(characters))
-    elif error_code == 1:
-        raise LookupError("Error with weird symbos" + "".join(characters))
-    elif error_code == 2:
-        raise ProgrammingError("Error with weird symbos" + "".join(characters))